backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlmodel==0.0.14
asyncpg==0.29.0
alembic==1.13.1
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
redis==5.0.1
celery==5.3.4
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
paddleocr==2.7.3
paddlepaddle==3.0.0
openai-whisper==20231117
torch==2.1.1
torchaudio==2.1.1
numpy==1.24.3
Pillow==10.1.0
python-dotenv==1.0.0